Refactor into reusable components
